Added project files to git
Added a staggered loader to load assets on startup.
Use the clear models pulse button on the geo_city component to delete models and then save project and quit. 
Hid the splash screen on startup
Implemented a naive view frustrum culling in SOPS (Very CPU heavy though and needs to be optimized)
FPS is worse when camera is moving because all the allocator SOPs need to cook
Possibly rework the allocator network